split off TestResamplerGrouper class
